Fixed Review rating from Integer to double and review to reviewText

+++ models/review.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Review = sequelize.define(
     "Review",
     {
-      rating: DataTypes.INTEGER,
+      rating: DataTypes.DOUBLE,
       review: DataTypes.TEXT
     },
     {}
